renamed from Codeforces Tools
